linebot-reminder/requirements.txt
Flask==2.0.3
line-bot-sdk==1.19.0
python-dotenv==0.19.2
dateparser==1.1.1